Re-factor and tidy up after adding migration lambda
