+++ src/migrateShards.ts
@@ -1,6 +1,6 @@
-import {ssmCommand} from './utils/ssmCommand'
-import {ClusterSizeCheckResponse, DefaultResponse} from './handlerResponses';
-import {ElasticsearchNode} from './utils/elasticsearch';
+import {ClusterSizeCheckResponse, DefaultResponse} from './utils/handlerResponses';
+import {getShardInfo, migrateShards} from './elasticsearch/elasticsearch';
+import {ElasticsearchNode, Move, ShardCopy} from './elasticsearch/types'
 
 export async function handler(event: ClusterSizeCheckResponse): Promise<DefaultResponse> {
     const oldestElasticsearchNode: ElasticsearchNode = event.oldestElasticsearchNode;
@@ -8,26 +8,6 @@ export async function handler(event: ClusterSizeCheckResponse): Promise<DefaultR
     return getShardInfo(oldestElasticsearchNode.ec2Instance.id)
         .then(response => moveInstructions(response, oldestElasticsearchNode, newestElasticsearchNode))
         .then(moves => migrateShards(moves, oldestElasticsearchNode));
-}
-
-function getShardInfo(instanceId: string): Promise<object> {
-    return ssmCommand(`curl localhost:9200/_cluster/state/routing_table`, instanceId)
-        .then(JSON.parse);
-}
-
-function migrateShards(moves: Move[], oldestElasticsearchNode: ElasticsearchNode): Promise<DefaultResponse> {
-    const commands: MoveCommand[] = moves.map(buildMoveCommand);
-    const rerouteCommand: RerouteCommand = {
-        commands
-    };
-    console.log(`Re-route command will be: ${JSON.stringify(rerouteCommand)}`);
-    const elasticsearchCommand = `curl -f -v -X POST "localhost:9200/_cluster/reroute" -H 'Content-Type: application/json' -d '${JSON.stringify(rerouteCommand)}'`;
-    return ssmCommand(elasticsearchCommand, oldestElasticsearchNode.ec2Instance.id)
-        .then(() => {
-                const response: DefaultResponse = {"oldestElasticsearchNode": oldestElasticsearchNode};
-                return response;
-            }
-        )
 }
 
 export function moveInstructions(response, oldNode: ElasticsearchNode, newNode: ElasticsearchNode): Move[] {
@@ -45,39 +25,6 @@ export function moveInstructions(response, oldNode: ElasticsearchNode, newNode: 
 
 function findShardCopiesOnNode(shardCopies: ShardCopy[], elasticsearchNode: ElasticsearchNode): ShardCopy[] {
     return shardCopies.filter(copy => copy.node === elasticsearchNode.nodeId);
-}
-
-interface ShardCopy {
-    node: string;
-    shard: number;
-    index: string;
-}
-
-interface RerouteCommand {
-    commands: MoveCommand[]
-}
-
-interface MoveCommand {
-    move: Move
-}
-
-export class Move {
-
-    index: string;
-    shard: number;
-    from_node: string;
-    to_node: string;
-
-    constructor(indexName, shardNumber, oldNode, newNode) {
-        this.index = indexName;
-        this.shard = shardNumber;
-        this.from_node = oldNode;
-        this.to_node = newNode;
-    }
-}
-
-function buildMoveCommand(move: Move): MoveCommand {
-    return { "move": move };
 }
